+++ test/commands/do/do_push_test.dart
@@ -67,7 +67,7 @@ void main() async {
     canCommit = MockCanCommit();
     doCommit = DoCommit(ggLog: ggLog, canCommit: canCommit);
     isPushed = IsPushed(ggLog: ggLog);
-    ggJson = File(join(dLocal.path, '.gg.json'));
+    ggJson = File(join(dLocal.path, '.gg', '.gg.json'));
 
     // Init pubspec.yaml
     await File(
@@ -215,7 +215,7 @@ void main() async {
                 isTrue,
               );
 
-              // Run ggDoPush should update .gg.json
+              // Run ggDoPush should update .gg/.gg.json
               final ggJsonBefore = await ggJson.readAsString();
               await doPush.exec(directory: dLocal, ggLog: ggLog);
               final ggJsonAfter = await ggJson.readAsString();