+++ lib/gg.dart
@@ -6,44 +6,39 @@
 
 library;
 
-export 'src/gg.dart';
-
+// Can
+export 'src/commands/can.dart';
+export 'src/commands/can/can_commit.dart';
+export 'src/commands/can/can_merge.dart';
+export 'src/commands/can/can_publish.dart';
+export 'src/commands/can/can_push.dart';
+export 'src/commands/can/can_upgrade.dart';
 // Checks
 export 'src/commands/check.dart';
 export 'src/commands/check/analyze.dart';
 export 'src/commands/check/format.dart';
 export 'src/commands/check/pana.dart';
-
-// Can
-export 'src/commands/can.dart';
-export 'src/commands/can/can_push.dart';
-export 'src/commands/can/can_commit.dart';
-export 'src/commands/can/can_publish.dart';
-export 'src/commands/can/can_upgrade.dart';
-export 'src/commands/can/can_merge.dart';
-
-// Do
-export 'src/commands/do.dart';
-export 'src/commands/do/do_push.dart';
-export 'src/commands/do/do_commit.dart';
-export 'src/commands/do/do_publish.dart';
-export 'src/commands/do/do_upgrade.dart';
-export 'src/commands/do/do_maintain.dart';
-export 'src/commands/do/do_merge.dart';
-
 // Did
 export 'src/commands/did.dart';
-export 'src/commands/did/did_push.dart';
 export 'src/commands/did/did_commit.dart';
-export 'src/commands/did/did_publish.dart';
-export 'src/commands/did/did_upgrade.dart';
 export 'src/commands/did/did_merge.dart';
-
-// Tools
-export 'src/tools/carriage_return.dart';
-export 'src/tools/command_cluster.dart';
-export 'src/tools/checks.dart';
-export 'src/tools/gg_state.dart';
-
+export 'src/commands/did/did_publish.dart';
+export 'src/commands/did/did_push.dart';
+export 'src/commands/did/did_upgrade.dart';
+// Do
+export 'src/commands/do.dart';
+export 'src/commands/do/do_commit.dart';
+export 'src/commands/do/do_maintain.dart';
+export 'src/commands/do/do_merge.dart';
+export 'src/commands/do/do_publish.dart';
+export 'src/commands/do/do_push.dart';
+export 'src/commands/do/do_upgrade.dart';
 // Info
 export 'src/commands/info.dart';
+export 'src/gg.dart';
+// Tools
+export 'src/tools/carriage_return.dart';
+export 'src/tools/checks.dart';
+export 'src/tools/command_cluster.dart';
+export 'src/tools/gg_state.dart';
+export 'src/tools/version_selector.dart';